# apispec.json
{
  "definitions": {
    "approved_dataset_reply": {
      "properties": {
        "dataset_id": {
          "type": "string"
        }, 
        "downvote": {
          "type": "string"
        }, 
        "name": {
          "type": "string"
        }, 
        "upvote": {
          "type": "string"
        }
      }
    }, 
    "dataset_reply": {
      "properties": {
        "category": {
          "type": "string"
        }, 
        "catogery_id": {
          "type": "string"
        }, 
        "dataset_id": {
          "type": "string"
        }, 
        "dataset_owner": {
          "type": "string"
        }, 
        "discription": {
          "type": "string"
        }, 
        "file_format": {
          "type": "string"
        }, 
        "file_patch": {
          "type": "string"
        }, 
        "last_update": {
          "type": "string"
        }, 
        "learning_tags": {
          "type": "string"
        }, 
        "link_to_api": {
          "type": "string"
        }, 
        "name": {
          "type": "string"
        }, 
        "search_tags": {
          "type": "string"
        }, 
        "status": {
          "type": "string"
        }, 
        "uploader": {
          "type": "string"
        }, 
        "visualisation_link": {
          "type": "string"
        }
      }
    }, 
...
    "get_votes_reply": {
      "properties": {
        "downvote": {
          "type": "integer"
        }, 
        "upvote": {
          "type": "integer"
        }
      }
    }, 
    "leader_boards_reply": {
      "properties": {
        "comment_desc": {
          "type": "string"
        }, 
        "first_name": {
          "type": "string"
        }, 
        "last_name": {
          "type": "string"
        }, 
        "rating": {
          "type": "string"
        }, 
        "time_stamp": {
          "type": "object"
        }
      }
    }, 
...
    "search_dataset_reply": {
      "properties": {
        "category": {
          "type": "string"
        }, 
        "catogery_id": {
          "type": "string"
        }, 
        "dataset_id": {
          "type": "string"
        }, 
        "dataset_owner": {
          "type": "string"
        }, 
        "discription": {
          "type": "string"
        }, 
        "file_format": {
          "type": "string"
        }, 
        "file_patch": {
          "type": "string"
        }, 
        "last_update": {
          "type": "string"
        }, 
        "learning_tags": {
          "type": "string"
        }, 
        "link_to_api": {
          "type": "string"
        }, 
        "name": {
          "type": "string"
        }, 
        "search_tags": {
          "type": "string"
        }, 
        "status": {
          "type": "string"
        }, 
        "uploader": {
          "type": "string"
        }, 
        "visualisation_link": {
          "type": "string"
        }
      }
    }, 
    "unapproved_dataset_reply": {
      "properties": {
        "category": {
          "type": "string"
        }, 
        "catogery_id": {
          "type": "string"
        }, 
        "dataset_id": {
          "type": "string"
        }, 
        "dataset_owner": {
          "type": "string"
        }, 
        "discription": {
          "type": "string"
        }, 
        "file_format": {
          "type": "string"
        }, 
        "file_patch": {
          "type": "string"
        }, 
        "last_update": {
          "type": "string"
        }, 
        "learning_tags": {
          "type": "string"
        }, 
        "link_to_api": {
          "type": "string"
        }, 
        "name": {
          "type": "string"
        }, 
        "search_tags": {
          "type": "string"
        }, 
        "status": {
          "type": "string"
        }, 
        "uploader": {
          "type": "string"
        }, 
        "visualisation_link": {
          "type": "string"
        }
      }
    }, 
...
  }, 
  "info": {
    "description": "", 
    "title": "The Learning Hub API"
  }, 
  "paths": {
    "/add_votes": {
      "post": {
        "description": "Call this api to get votes<br/>", 
        "parameters": [
          {
            "description": "The dataset ID", 
            "in": "body", 
            "name": "id", 
            "required": true, 
            "schema": {
              "properties": {
                "downvote": {
                  "type": "string"
                }, 
                "id": {
                  "type": "integer"
                }, 
                "upvote": {
                  "type": "string"
                }
              }
            }, 
            "type": "string"
          }
        ], 
        "responses": {
          "200": {
            "description": "success"
          }, 
          "400": {
            "description": "fail"
          }
        }, 
        "summary": "This is the get votes API", 
        "tags": [
          "API to get votes"
        ]
      }
    }, 
    "/approve_dataset": {
      "post": {
        "description": "Call this api to approve dataset<br/>", 
        "parameters": [
          {
            "description": "The approve data", 
            "in": "body", 
            "name": "data", 
            "required": true, 
            "schema": {
              "properties": {
                "action": {
                  "default": "approve", 
                  "type": "string"
                }, 
                "id": {
                  "type": "integer"
                }
              }
            }, 
            "type": "string"
          }
        ], 
        "responses": {
          "200": {
            "description": "success"
          }, 
          "400": {
            "description": "fail"
          }
        }, 
        "summary": "This is the approve dataset API", 
        "tags": [
          "API to approve dataset"
        ]
      }
    }, 
    "/approved_dataset": {
      "get": {
        "description": "Call this api to get approved dataset<br/>", 
        "responses": {
          "200": {
            "description": "Success", 
            "schema": {
              "$ref": "#/definitions/approved_dataset_reply"
            }
          }, 
          "400": {
            "description": "Error"
          }
        }, 
        "summary": "This is the get approved dataset API", 
        "tags": [
          "API to get approved dataset"
        ]
      }
    }, 
    "/comments/{id}": {
      "get": {
        "description": "Call this api to get or post comments<br/>", 
        "parameters": [
          {
            "description": "The data ID", 
            "in": "path", 
            "name": "data_id", 
            "required": true, 
            "type": "string"
          }
        ], 
        "responses": {
          "200": {
            "description": "Success", 
            "schema": {
              "$ref": "#/definitions/leader_boards_reply"
            }
          }, 
          "500": {
            "description": "Error"
          }
        }, 
        "summary": "This is the comments API", 
        "tags": [
          "API to get or post comments"
        ]
      }, 
      "post": {
        "description": "Call this api to get or post comments<br/>", 
        "parameters": [
          {
            "description": "The data ID", 
            "in": "path", 
            "name": "data_id", 
            "required": true, 
            "type": "string"
          }, 
          {
            "description": "The comment desc", 
            "in": "body", 
            "name": "name", 
            "required": true, 
            "schema": {
              "properties": {
                "name": {
                  "type": "string"
                }
              }
            }, 
            "type": "string"
          }
        ], 
        "responses": {
          "200": {
            "description": "Success", 
            "schema": {
              "$ref": "#/definitions/leader_boards_reply"
            }
          }, 
          "500": {
            "description": "Error"
          }
        }, 
        "summary": "This is the comments API", 
        "tags": [
          "API to get or post comments"
        ]
      }
    }, 
    "/dataset_API/{id}": {
      "get": {
        "description": "Call this api passing a dataset id and get back its content<br/>", 
        "parameters": [
          {
            "description": "The dataset ID", 
            "in": "path", 
            "name": "id", 
            "required": true, 
            "type": "integer"
          }
        ], 
        "responses": {
          "200": {
            "description": "Success", 
            "schema": {
              "$ref": "#/definitions/dataset_reply"
            }
          }, 
          "500": {
            "description": "Error"
          }
        }, 
        "summary": "This is the get dataset API", 
        "tags": [
          "API to get dataset"
        ]
      }
    }, 
...
    "/get_votes/{id}": {
      "get": {
        "description": "Call this api to get votes<br/>", 
        "parameters": [
          {
            "description": "The dataset ID", 
            "in": "path", 
            "name": "id", 
            "required": true, 
            "type": "string"
          }
        ], 
        "responses": {
          "200": {
            "description": "success", 
            "schema": {
              "$ref": "#/definitions/get_votes_reply"
            }
          }, 
          "400": {
            "description": "fail"
          }
        }, 
        "summary": "This is the get votes API", 
        "tags": [
          "API to get votes"
        ]
      }
    }, 
...
    "/search_dataset/{search_term}": {
      "get": {
        "description": "Call this api to search dataset<br/>", 
        "parameters": [
          {
            "description": "The search term", 
            "in": "path", 
            "name": "search_term", 
            "required": true, 
            "type": "string"
          }
        ], 
        "responses": {
          "200": {
            "description": "success", 
            "schema": {
              "$ref": "#/definitions/search_dataset_reply"
            }
          }, 
          "400": {
            "description": "fail"
          }, 
          "500": {
            "description": null
          }
        }, 
        "summary": "This is the search dataset API", 
        "tags": [
          "API to search dataset"
        ]
      }
    }, 
...
    "/unapproved_dataset": {
      "get": {
        "description": "Call this api to get unapproved dataset<br/>", 
        "responses": {
          "200": {
            "description": "Success", 
            "schema": {
              "$ref": "#/definitions/unapproved_dataset_reply"
            }
          }, 
          "400": {
            "description": "Error"
          }
        }, 
        "summary": "This is the get unapproved dataset API", 
        "tags": [
          "API to get unapproved dataset"
        ]
      }
    }, 
...
  }, 
  "swagger": "2.0"
}